refactor Name
